Split ResultPopup
